+++ .eslint-doc-generatorrc.js
@@ -11,6 +11,13 @@ module.exports = {
     'rules-recommended',
     'tests',
     'tests-recommended',
+    'flat/recommended',
+    'flat/all',
+    'flat/all-type-checked',
+    'flat/rules',
+    'flat/rules-recommended',
+    'flat/tests',
+    'flat/tests-recommended',
   ],
   postprocess: async (content, path) =>
     prettier.format(content, {