+++ configs/tests.js
@@ -1,13 +1,11 @@
 /**
  * @fileoverview the `tests` config for `eslint.config.js`
+ * @deprecated use 'flat/tests' instead
  * @author 唯然<weiran.zsd@outlook.com>
  */
 
 'use strict';
 
-const mod = require('../lib/index.js');
+const plugin = require('../lib/index.js');
 
-module.exports = {
-  plugins: { 'eslint-plugin': mod },
-  rules: mod.configs.tests.rules,
-};
+module.exports = plugin.configs['flat/tests'];