+++ tests/lib/rules/require-meta-schema.js
@@ -5,13 +5,15 @@
 // ------------------------------------------------------------------------------
 
 const rule = require('../../../lib/rules/require-meta-schema');
-const RuleTester = require('eslint').RuleTester;
+const RuleTester = require('../eslint-rule-tester').RuleTester;
 
 // ------------------------------------------------------------------------------
 // Tests
 // ------------------------------------------------------------------------------
 
-const ruleTester = new RuleTester({ parserOptions: { ecmaVersion: 9 } });
+const ruleTester = new RuleTester({
+  languageOptions: { sourceType: 'commonjs' },
+});
 ruleTester.run('require-meta-schema', rule, {
   valid: [
     `
@@ -61,7 +63,7 @@ ruleTester.run('require-meta-schema', rule, {
           create(context) { const options = context.options; }
         };
       `,
-      parserOptions: { sourceType: 'module' },
+      languageOptions: { sourceType: 'module' },
     },
     // Variable schema with array value.
     `
@@ -254,7 +256,7 @@ schema: []
         };
       `,
       output: null,
-      parserOptions: { sourceType: 'module' },
+      languageOptions: { sourceType: 'module' },
       errors: [
         {
           messageId: 'missing',