+++ tests/lib/rules/test-case-shorthand-strings.js
@@ -10,7 +10,7 @@
 // ------------------------------------------------------------------------------
 
 const rule = require('../../../lib/rules/test-case-shorthand-strings');
-const RuleTester = require('eslint').RuleTester;
+const RuleTester = require('../eslint-rule-tester').RuleTester;
 
 /**
  * Returns the code for some valid test cases
@@ -40,7 +40,9 @@ const UNEXPECTED_SHORTHAND_ERROR = {
 // Tests
 // ------------------------------------------------------------------------------
 
-const ruleTester = new RuleTester({ parserOptions: { ecmaVersion: 6 } });
+const ruleTester = new RuleTester({
+  languageOptions: { sourceType: 'commonjs' },
+});
 ruleTester.run('test-case-shorthand-strings', rule, {
   valid: [
     // default (as-needed)